dev-back/requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-cors-headers==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.9.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
mysqlclient==1.4.2.post1
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
unicodecsv==0.14.1
djangorestframework-csv==2.1.0
djangorestframework-camel-case